Potential fix to stacking TimedActions in queue
