+++ media/lua/client/AutoGate/ISAutoGateTimedActions.lua
@@ -20,9 +20,97 @@ local ISAutoGateTooltip = require "AutoGate/ISUI/ISAutoGateTooltip"
 local ISAutoGateInstallAction = require "AutoGate/TimedActions/ISAutoGateInstallAction"
 local ISAutoGateInteractAction = require "AutoGate/TimedActions/ISAutoGateInteractAction"
 local ISAutoGateControllerAction = require "AutoGate/TimedActions/ISAutoGateControllerAction"
+local ISAutoGateClearQueueAction = require "AutoGate/TimedActions/ISAutoGateClearQueueAction"
 
 local BlowtorchUtils = ISBlacksmithMenu
 
+
+------------------ Functions related to TimedActions Checks ------------------
+local function comparatorDrainableUsesInt(item1, item2)
+    return item1:getDrainableUsesInt() - item2:getDrainableUsesInt()
+end
+---Get the best Welding Rods inside a inventory container
+---@param container ItemContainer Usually the player inventory
+---@return DrainableComboItem WeldingRods with most uses left
+function ISAutoGateTimedActions.getWeldingRodsWithMostUses(container)
+    return container:getBestTypeEvalRecurse("Base.WeldingRods", comparatorDrainableUsesInt)
+end
+
+---Check if the item still have uses left
+---@param inventoryItem InventoryItem Item to be checked
+---@param itemType String Type of the item
+---@return boolean True if has uses left, false if not
+function ISAutoGateTimedActions.hasDeltaLeft(inventoryItem, itemType)
+    if inventoryItem:getType() == itemType then
+        if inventoryItem:getDelta() > 0 then
+            return true
+        end
+    end
+    return false
+end
+
+---Equips items for TimedAction
+---@param player IsoPlayer Player doing the action
+---@return DrainableComboItem, DrainableComboItem
+function ISAutoGateTimedActions.checkAndEquipInstallItems(player)
+    local playerInventory = player:getInventory()
+
+    --Checking if equipped items are valid
+    local equippedPrimary = player:getPrimaryHandItem()
+    local alreadyEquippedPrimary = false
+    if instanceof(equippedPrimary, "DrainableComboItem") then
+        alreadyEquippedPrimary = ISAutoGateTimedActions.hasDeltaLeft(equippedPrimary, "BlowTorch")
+    end
+    local equippedSecondary = player:getSecondaryHandItem()
+    local alreadyEquippedSecondary = false
+    if instanceof(equippedSecondary, "DrainableComboItem") then
+        alreadyEquippedSecondary = ISAutoGateTimedActions.hasDeltaLeft(equippedSecondary, "WeldingRods")
+    end
+    --Setting correct items
+    local blowtorch = BlowtorchUtils.getBlowTorchWithMostUses(playerInventory)
+    if alreadyEquippedPrimary then
+        blowtorch = equippedPrimary
+    end
+    local weldingrods = ISAutoGateTimedActions.getWeldingRodsWithMostUses(playerInventory)
+    if alreadyEquippedSecondary then
+        weldingrods = equippedSecondary
+    end
+    local weldingmask = playerInventory:getItemFromTypeRecurse("WeldingMask")
+
+    ISInventoryPaneContextMenu.transferIfNeeded(player, blowtorch)
+    ISInventoryPaneContextMenu.transferIfNeeded(player, weldingrods)
+    ISInventoryPaneContextMenu.transferIfNeeded(player, weldingmask)
+    luautils.equipItems(player, blowtorch, weldingrods)
+    ISInventoryPaneContextMenu.wearItem(weldingmask, player:getPlayerNum())
+    return blowtorch, weldingrods
+end
+
+---Get the items for a given interaction and equips them
+---@param player IsoPlayer Player to do the action
+---@param gateOrController string If the action is on a gate/controller
+---@return InventoryItem, table<number,InventoryItem> Return checked item and it's container instance
+function ISAutoGateTimedActions.checkInteractItem(player, gateOrController)
+    local playerInventory = player:getInventory()
+    local returnToContainer = {}
+    local screwdriver = playerInventory:getItemFromTypeRecurse("Screwdriver")
+    local wrench = playerInventory:getItemFromTypeRecurse("Wrench")
+    if gateOrController == "controller" and screwdriver then
+        table.insert(returnToContainer, screwdriver)
+        ISInventoryPaneContextMenu.transferIfNeeded(player, screwdriver)
+        return screwdriver, returnToContainer
+    elseif gateOrController == "gate" and wrench then
+        table.insert(returnToContainer, wrench)
+        ISInventoryPaneContextMenu.transferIfNeeded(player, wrench)
+        return wrench, returnToContainer
+    elseif gateOrController == "both" and wrench and screwdriver then
+        table.insert(returnToContainer, wrench)
+        table.insert(returnToContainer, screwdriver)
+        ISInventoryPaneContextMenu.transferIfNeeded(player, wrench)
+        ISInventoryPaneContextMenu.transferIfNeeded(player, screwdriver)
+        return wrench, screwdriver, returnToContainer
+    end
+    return nil
+end
 
 ------------------ Functions related to gate installation ------------------
 
@@ -39,7 +127,7 @@ function ISAutoGateTimedActions.addOptionInstallAutomaticMotor(player, context, 
     local components = playerInventory:getCountTypeRecurse("GateComponents")
     local blowtorch = BlowtorchUtils.getBlowTorchWithMostUses(playerInventory)
     local blowtorchUses = 0
-    local weldingrods = ISAutoGateUtils.getWeldingRodsWithMostUses(playerInventory)
+    local weldingrods = ISAutoGateTimedActions.getWeldingRodsWithMostUses(playerInventory)
     local weldingrodsUses = 0
     local weldingmask = playerInventory:getCountTypeRecurse("WeldingMask")
     ------------------ Running checks ------------------
@@ -67,8 +155,9 @@ function ISAutoGateTimedActions.queueInstallAutomaticGateMotor(player, gate)
     local doorSquare = gateOppositeSquare:DistTo(playerSquare) < gateSquare:DistTo(playerSquare) and gateOppositeSquare or gateSquare
 
     ISTimedActionQueue.add(ISWalkToTimedAction:new(player, doorSquare))
-    local blowtorch, weldingrods = ISAutoGateUtils.checkAndEquipInstallItems(player)
+    local blowtorch, weldingrods = ISAutoGateTimedActions.checkAndEquipInstallItems(player)
     ISTimedActionQueue.add(ISAutoGateInstallAction:new(player, gateCornerObject, blowtorch, weldingrods))
+    ISTimedActionQueue.add(ISAutoGateClearQueueAction:new(player))
 end
 
 ------------------ Functions related to gate and controller interactions ------------------
@@ -86,10 +175,11 @@ function ISAutoGateTimedActions.connectControllerToGate(player, emptyController,
     local doorSquare = (gateOppositeSquare:DistTo(playerSquare) < gateSquare:DistTo(playerSquare)) and gateOppositeSquare or gateSquare
     ISTimedActionQueue.add(ISWalkToTimedAction:new(player, doorSquare))
 
-    local wrench, screwdriver, returnItems = ISAutoGateUtils.checkInteractItem(player, "both")
+    local wrench, screwdriver, returnItems = ISAutoGateTimedActions.checkInteractItem(player, "both")
     ISTimedActionQueue.add(ISAutoGateInteractAction:new(player, gate, wrench, "connect"))
     ISTimedActionQueue.add(ISAutoGateControllerAction:new(player, screwdriver, emptyController, "connect", nil, gate))
     ISCraftingUI.ReturnItemsToOriginalContainer(player, returnItems)
+    ISTimedActionQueue.add(ISAutoGateClearQueueAction:new(player))
 end
 
 ---Resets a automatic gate frequency
@@ -104,9 +194,10 @@ function ISAutoGateTimedActions.resetGate(gate, player)
     local doorSquare = (gateOppositeSquare:DistTo(playerSquare) < gateSquare:DistTo(playerSquare)) and gateOppositeSquare or gateSquare
     ISTimedActionQueue.add(ISWalkToTimedAction:new(player, doorSquare))
 
-    local wrench, returnItems = ISAutoGateUtils.checkInteractItem(player, "gate")
+    local wrench, returnItems = ISAutoGateTimedActions.checkInteractItem(player, "gate")
     ISTimedActionQueue.add(ISAutoGateInteractAction:new(player, gate, wrench, "reset"))
     ISCraftingUI.ReturnItemsToOriginalContainer(player, returnItems)
+    ISTimedActionQueue.add(ISAutoGateClearQueueAction:new(player))
 end
 
 ---Copies the frequency from a connected controller to another
@@ -114,24 +205,22 @@ end
 ---@param fromConnectedController InventoryItem Controller with a connection
 ---@param toEmptyController InventoryItem Controller without a connection
 function ISAutoGateTimedActions.copyControllerToAnother(player, fromConnectedController, toEmptyController)
-    local screwdriver, returnItems = ISAutoGateUtils.checkInteractItem(player, "controller")
+    local screwdriver, returnItems = ISAutoGateTimedActions.checkInteractItem(player, "controller")
     ISTimedActionQueue.add(ISAutoGateControllerAction:new(player, screwdriver, fromConnectedController, "copyStart", toEmptyController))
     ISTimedActionQueue.add(ISAutoGateControllerAction:new(player, screwdriver, toEmptyController, "copyFinish", fromConnectedController))
     ISCraftingUI.ReturnItemsToOriginalContainer(player, returnItems)
+    ISTimedActionQueue.add(ISAutoGateClearQueueAction:new(player))
 end
 
 ---Disconnects a controller from a gate
 ---@param player IsoPlayer Player
 ---@param connectedController InventoryItem Already connected controller
 function ISAutoGateTimedActions.disconnectController(player, connectedController)
-    local screwdriver, returnItems = ISAutoGateUtils.checkInteractItem(player, "controller")
+    local screwdriver, returnItems = ISAutoGateTimedActions.checkInteractItem(player, "controller")
     ISTimedActionQueue.add(ISAutoGateControllerAction:new(player, screwdriver, connectedController, "disconnect"))
     ISCraftingUI.ReturnItemsToOriginalContainer(player, returnItems)
+    ISTimedActionQueue.add(ISAutoGateClearQueueAction:new(player))
 end
-
-
-
-
 
 
 ------------------ Returning file for 'require' ------------------
