+++ media/lua/shared/AutoGate/ISAutoGateUtils.lua
@@ -175,95 +175,6 @@ end
 
 --[[**********************************************************************************]]--
 
------------------- Functions related to TimedActions ------------------
-local function comparatorDrainableUsesInt(item1, item2)
-    return item1:getDrainableUsesInt() - item2:getDrainableUsesInt()
-end
----Get the best Welding Rods inside a inventory container
----@param container ItemContainer Usually the player inventory
----@return DrainableComboItem WeldingRods with most uses left
-function ISAutoGateUtils.getWeldingRodsWithMostUses(container)
-    return container:getBestTypeEvalRecurse("Base.WeldingRods", comparatorDrainableUsesInt)
-end
-
----Check if the item still have uses left
----@param inventoryItem InventoryItem Item to be checked
----@param itemType String Type of the item
----@return boolean True if has uses left, false if not
-function ISAutoGateUtils.hasDeltaLeft(inventoryItem, itemType)
-	if inventoryItem:getType() == itemType then
-		if inventoryItem:getDelta() > 0 then
-			return true
-		end
-	end
-	return false
-end
-
----Equips items for TimedAction
----@param player IsoPlayer Player doing the action
----@return DrainableComboItem, DrainableComboItem
-function ISAutoGateUtils.checkAndEquipInstallItems(player)
-	local playerInventory = player:getInventory()
-
-	--Checking if equipped items are valid
-	local equippedPrimary = player:getPrimaryHandItem()
-	local alreadyEquippedPrimary = false
-	if instanceof(equippedPrimary, "DrainableComboItem") then
-		alreadyEquippedPrimary = ISAutoGateUtils.hasDeltaLeft(equippedPrimary, "BlowTorch")
-	end	
-	local equippedSecondary = player:getSecondaryHandItem()
-	local alreadyEquippedSecondary = false
-	if instanceof(equippedSecondary, "DrainableComboItem") then
-		alreadyEquippedSecondary = ISAutoGateUtils.hasDeltaLeft(equippedSecondary, "WeldingRods")
-	end
-	--Setting correct items
-	local blowtorch = ISBlacksmithMenu.getBlowTorchWithMostUses(playerInventory)
-	if alreadyEquippedPrimary then
-		blowtorch = equippedPrimary
-	end
-	local weldingrods = ISAutoGateUtils.getWeldingRodsWithMostUses(playerInventory)
-	if alreadyEquippedSecondary then
-		weldingrods = equippedSecondary
-	end
-	local weldingmask = playerInventory:getItemFromTypeRecurse("WeldingMask")
-
-	ISInventoryPaneContextMenu.transferIfNeeded(player, blowtorch)
-	ISInventoryPaneContextMenu.transferIfNeeded(player, weldingrods)
-	ISInventoryPaneContextMenu.transferIfNeeded(player, weldingmask)
-	luautils.equipItems(player, blowtorch, weldingrods)
-	ISInventoryPaneContextMenu.wearItem(weldingmask, player:getPlayerNum())
-	return blowtorch, weldingrods
-end
-
----Get the items for a given interaction and equips them
----@param player IsoPlayer Player to do the action
----@param gateOrController string If the action is on a gate/controller
----@return InventoryItem, table<number,InventoryItem> Return checked item and it's container instance
-function ISAutoGateUtils.checkInteractItem(player, gateOrController)
-	local playerInventory = player:getInventory()
-	local returnToContainer = {}
-	local screwdriver = playerInventory:getItemFromTypeRecurse("Screwdriver")
-	local wrench = playerInventory:getItemFromTypeRecurse("Wrench")
-	if gateOrController == "controller" and screwdriver then
-		table.insert(returnToContainer, screwdriver)
-		ISInventoryPaneContextMenu.transferIfNeeded(player, screwdriver)
-		return screwdriver, returnToContainer
-	elseif gateOrController == "gate" and wrench then
-		table.insert(returnToContainer, wrench)
-		ISInventoryPaneContextMenu.transferIfNeeded(player, wrench)
-		return wrench, returnToContainer
-	elseif gateOrController == "both" and wrench and screwdriver then
-		table.insert(returnToContainer, wrench)
-		table.insert(returnToContainer, screwdriver)
-		ISInventoryPaneContextMenu.transferIfNeeded(player, wrench)
-		ISInventoryPaneContextMenu.transferIfNeeded(player, screwdriver)
-		return wrench, screwdriver, returnToContainer
-	end
-	return nil
-end
-
---[[**********************************************************************************]]--
-
 ------------------ Functions related to getting a specific item from containers (gate's or player's inventory) ------------------
 ---Simple predicate function to get a battery with charge
 ---@param batteriesItems table<number, InventoryItem>
